For #61: review fixes.
